committed

+++ lib/firebase/FirebaseStorageUtil.dart
@@ -40,8 +40,6 @@ class FirebaseStorageUtil {
     if (null == chat.thumbnailPath || "" == chat.thumbnailPath) {
       chat.thumbnailPath = await Thumbnails.getThumbnail(
           videoFile: chat.localPath, imageType: ThumbFormat.PNG, quality: 11);
-
-      //print('Thumbnail generated at ' + chat.thumbnailPath);
     }
 
     return chat.thumbnailPath;
@@ -58,7 +56,6 @@ class FirebaseStorageUtil {
   }
 
   Future<bool> checkIfFileExists(String filePath) async {
-    //print('checking file exists ' + filePath);
     return await File(filePath).exists();
   }
 
@@ -90,12 +87,6 @@ class FirebaseStorageUtil {
       ProgressBloc().addToInProgressMap(user.id);
       String fileName = user.id + '.' + user.photoUrl.split('.').last;
       String path = 'UserPhotos/' + fileName;
-      /*print('in _addUserPhotoToFB ' +
-          user.photoUrl +
-          ' ' +
-          user.id +
-          ' file name ' +
-          fileName);*/
       File compressedImage;
       if (compressMedia) {
         compressedImage = await FlutterNativeImage.compressImage(user.photoUrl,
@@ -132,7 +123,6 @@ class FirebaseStorageUtil {
 
   Future<void> addFileToFirebaseStorage(
       ChatModel chat, bool compressMedia) async {
-    //print('calling addFileToFirebaseStorage id ' + chat.id.toString());
     if (!ProgressBloc().isUploadInProgress(chat.id.toString())) {
       StorageUploadTask task;
       ProgressBloc().addToInProgressMap(chat.id.toString());
@@ -147,7 +137,6 @@ class FirebaseStorageUtil {
           Utils().getChatCollectionId(chat.fromUserId, chat.toUserId) +
           '/' +
           fileName;
-      //print('path created ' + path);
       final StorageReference storageReference = _ref.child(path);
       if (compressMedia) {
         File compressedImage = await FlutterNativeImage.compressImage(
@@ -164,7 +153,6 @@ class FirebaseStorageUtil {
 
       task.events.listen((data) async {
         if (data.type == StorageTaskEventType.failure) {
-          //print('error while uploading file');
           ProgressBloc().addToProgressController(
               ProgressModel(chat.id.toString(), ProgressModel.ERR));
           ProgressBloc().removeFromInProgressMap(chat.id.toString());
@@ -175,11 +163,7 @@ class FirebaseStorageUtil {
           if (filetype == 'UserMedia') {
             await addThumbnailToStorage(chat);
           }
-
-          /*print('closing progress controller after adding end for id ' +
-              chat.id.toString());*/
-
-          chat.firebaseStorage = path;
+    chat.firebaseStorage = path;
           await Firebase()
               .addUpdateChat(chat, Firebase.CHAT_COL_COMPLETE, true);
           ProgressBloc().addToProgressController(
@@ -198,18 +182,11 @@ class FirebaseStorageUtil {
         if (chat.localPath == null || "" == chat.localPath) {
           chat.localPath = dirPath + '/' + chat.firebaseStorage.split("/").last;
         }
-
-        //print('start checking file exists for ' + chat.id.toString());
         bool isFileExists =
             await checkIfFileExists(chat.localPath).catchError((e) {
           throw Exception(e);
         });
-        /*print('file exists ' +
-            isFileExists.toString() +
-            ' for chat id ' +
-            chat.id.toString());*/
         if (!isFileExists) {
-          //print('downloading file from network path ' + chat.firebaseStorage);
           final StorageReference storageReference =
               _ref.child(chat.firebaseStorage);
           StorageFileDownloadTask task =
@@ -223,7 +200,6 @@ class FirebaseStorageUtil {
         return File(chat.localPath);
       });
     } on Exception catch (e) {
-      //print('got error while download ' + e.toString());
       throw Exception();
     }
   }
@@ -261,7 +237,6 @@ class FirebaseStorageUtil {
         }
       });
     } else {
-      //print('db path for upload ' + localDbPath + ' does not exists');
     }
   }
 
@@ -269,17 +244,20 @@ class FirebaseStorageUtil {
       ChatDeleteModel chatDeleteModel) async {
     String fbPath = chatDeleteModel.fbStoragePath;
     String localPath = chatDeleteModel.localPath;
-    String thumbnailPath = chatDeleteModel.thumbnailPath;
+    
     _checkPathInFBStorageAndDelete(fbPath);
-
+      if(chatDeleteModel.thumbnailPath!=null && ''!=chatDeleteModel.thumbnailPath) {
+        String thumbnailPath = _getThumbnailPath(chatDeleteModel);
       _checkPathInFBStorageAndDelete(thumbnailPath);
     
+      }
+     
     _checkPathInLocalAndDelete(localPath);
   }
 
   Future<bool> _checkPathInFBStorageAndDelete(String path) async {
     if (null != path && '' != path.trim()) {
-      try {
+      try {        
         StorageReference _reference = _ref.child(path);
         StorageMetadata storageMetadata = await _reference.getMetadata();
         if (storageMetadata != null && storageMetadata.sizeBytes > 0) {
@@ -297,14 +275,23 @@ class FirebaseStorageUtil {
 
   Future<bool> _checkPathInLocalAndDelete(String path) async {
     if (null != path && '' != path.trim()) {
-      if (await checkIfFileExists(path)) {
+      if (await checkIfFileExists(path) && path.contains(PathConstants.CHATAPP_MEDIA)) {
         File file = File(path);
         file.deleteSync();
         print('deleted file from local ' + path);
         return true;
+      }else{
+        print('not deleting local path '+path);
       }
     }
     return false;
+  }
+
+  String _getThumbnailPath(ChatDeleteModel deleteModel) {
+       ChatModel chat = deleteModel.chat;
+       String path = 'Thumbnail/'+Utils().getChatCollectionId(chat.fromUserId, chat.toUserId)+'/'
+       +chat.id.toString()+'-thumb.png';
+       return path;
   }
 
   Future<bool> downloadLocalDB(String path) async {
